output_connectors/postgresql/requirements.txt
psycopg2-binary==2.8.4
ubiops-connector==1.0.2
